+++ src/main/java/net/geforcemods/securitycraft/datagen/BlockTagGenerator.java
@@ -934,7 +934,8 @@ public class BlockTagGenerator extends BlockTagsProvider {
 				SCContent.REINFORCED_WEATHERED_COPPER_BULB.get(),
 				SCContent.REINFORCED_OXIDIZED_COPPER_BULB.get(),
 				SCContent.REINFORCED_STONE_BUTTON.get(),
-				SCContent.REINFORCED_POLISHED_BLACKSTONE_BUTTON.get())
+				SCContent.REINFORCED_POLISHED_BLACKSTONE_BUTTON.get(),
+				SCContent.REINFORCED_LIGHTNING_ROD.get())
 		.addTag(SCTags.Blocks.REINFORCED_WALLS)
 		.addTag(SCTags.Blocks.REINFORCED_CAULDRONS);
 		tag(BlockTags.MINEABLE_WITH_SHOVEL).add(
@@ -1023,7 +1024,8 @@ public class BlockTagGenerator extends BlockTagsProvider {
 				SCContent.REINFORCED_COPPER_BULB.get(),
 				SCContent.REINFORCED_EXPOSED_COPPER_BULB.get(),
 				SCContent.REINFORCED_WEATHERED_COPPER_BULB.get(),
-				SCContent.REINFORCED_OXIDIZED_COPPER_BULB.get());
+				SCContent.REINFORCED_OXIDIZED_COPPER_BULB.get(),
+				SCContent.REINFORCED_LIGHTNING_ROD.get());
 	}
 
 	@Override